Restructure scheduling code to remove import cycles
This is not a fully coherent restructuring, as the scheduling code still
requires refactoring.
