+++ lib/scheduled-jobs.ts
@@ -1,7 +1,28 @@
-// eslint-disable-next-line import/no-cycle
-import Job from "./Job"
+import * as scheduler from "node-schedule"
+import * as hubot from "hubot"
 
-export type ScheduledJob = Job
+export type MessageMetadata = {
+  threadId?: string
+  messageId: string
+  lastUrl?: string
+}
+
+export type JobUser = Pick<hubot.User, "id" | "name"> & { room: string }
+
+export interface ScheduledJob {
+  id: string
+  pattern: string
+  room: string
+  message: string
+  metadata: MessageMetadata
+  remindInThread: boolean
+  user: JobUser
+  job?: scheduler.Job
+
+  isCron(): boolean
+  serialize(): readonly [string, JobUser, string, MessageMetadata, boolean]
+  cancel(): void
+}
 
 export type ScheduledJobMap = {
   [jobId: string]: ScheduledJob
